Add authenticator debug to entry gateway config
